requirements.txt
streamlit
pandas
plotly_express
matplotlib
seaborn
